Trust gloox on SSL certificates validation
I checked the Gloox sources and it already validates the certificate
validity dates as well as the server name. There is nothing more we need
to validate. We may want better handling (explicit error on invalid SSL,
allowing to ignore certificate problems, allow the user to set up their
own certs, etc), but that can be tracked in separate issues.

Fixes #30